Make big sweeping major breaking changes
After lots of usage, I have some insights about the problem space that
make me realize that the current implementation leaves some to be
desired. While the ESLint and Prettier combo are great, it's hard to
apply CSpell or TypeScript to the majority of 'normal' codebases. The
CSpell dictionary is tiny, and TypeScript doesn't understand that these
two things are the same:

```javascript
// Completely understood by TypeScript.
typeof s === 'string

// Completely opaque to TypeScript.
((x) => typeof x === 'string')(x)
```

The JavaScript codebases that I've been working with make heavy use of
utilities like `isString()`, which means that TypeScript reports an
unfortunate number of false positives that would require removing lots
of utility classes. For these reasons, this commit removes CSpell and
TypeScript.

This commit also removes Dependency-Check: it's a wonderful project, and
I still think this is critical, but it's minimalist aesthetic means that
we've had to add lots of `--ignore` garbage to `npm test`. There's a
similar project, Depcheck, which seems to be a less-minimalist approach
to the problem. My experience so far has been positive, and I'm
interested in continuing my experience with Depcheck.

This also introduces a `common-good both` command, which runs `fix` and
then `check`, and brings a Markdown plugin to ESLint so that our readmes
can get linted too! I'm interested in adding a TypeScript plugin as
well, but I'm losing steam for the day and want to push what I've got.