Add -staged commands and git pre-commit hook
Running all of the linters on all of the files is resource-intensive and
not very fun. In a perfect world, software engineers should only have to
worry about checking their staged files and they should be able to trust
that the rest of the files are already linted.

Using the -staged suffix runs all fixers and testers (with the exception
of depcheck) on only the staged files, and any mutations (e.g. Prettier
formatting) are also added to the staged file so that it can be
committed.